Move Configuration up
The Configuration class is not bound to the Console so this can be moved up by a namespace.
